get user profile webassembly working
