+++ .vscode/launch.json
@@ -6,6 +6,12 @@
     "configurations": [
         {
             "type": "node",
+            "request": "launch",
+            "name": "Launch Program",
+            "program": "${workspaceFolder}/server.js"
+        },
+        {
+            "type": "node",
             "request": "attach",
             "name": "Attach",
             "port": 5757
