--- experimental/hello.html
@@ -1,38 +1,0 @@
-<!doctypehtml><html lang=en-us><head><meta charset=utf-8><meta content="text/html; charset=utf-8"http-equiv=Content-Type><title>Emscripten-Generated Code</title></head><body>
-<input type="file" id="thisFile">
-<img src="" id="resultImage">
-<div id="progress"></div>
-
-<script>
-var fileInput = document.getElementById('thisFile');
-fileInput.addEventListener('change',function(){
-	if(this.files[0]){
-		var fr = new FileReader();
-		fr.onload = function () {
-			//todo: put this in a web worker
-			var imageProcesser = new Promise(function(resolve,reject){
-				var progressShower = Module.cwrap('getResizeProgress', 'number');
-				var progressWatcher = setInterval(function(){document.getElementById('progress').innerHTML = progressShower()+"%";},500);
-			
-				var data = new Uint8Array(fr.result);
-
-				Module['FS_createDataFile']('/', 'image', data, true, true, true);
-				Module.ccall('profileImage', null, [], null);
-
-				fileInput.value = '';
-				var imageResult = FS.readFile("output.png",{encoding:"binary"});
-				var blob = new Blob([imageResult], {'type': 'image/png'});
-				var url = URL.createObjectURL(blob);
-				clearInterval(progressWatcher);
-				resolve(url);
-			};
-			imageProcesser.then(function(url){
-				document.getElementById('resultImage').src = url;
-			}
-		})
-		fr.readAsArrayBuffer(this.files[0]);
-	}
-})
-</script>
-
-<script async src="hello.js"></script></body></html>
