+++ webassemblycomps/stb_image_resize.h
