+++ webassemblycomps/stb_image_write.h
